Move Model.to_json() to sphinxext.bokeh_model
